Refactor classes out from node.go into a new raft_server.go file.
